newFormV2/requirements.txt
Django==1.8
django-annoying==0.10.3
django-appconf==1.0.2
django-bower==5.2.0
django-user-accounts==2.0.3
funcsigs==1.0.2
icalendar==3.11.6
Markdown==2.6.8
mock==2.0.0
MySQL-python==1.2.5
pbr==4.0.4
psycopg2==2.7.1
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0